backend/db.py
# db.py
import mysql.connector

def get_connection():
    return mysql.connector.connect(
        host="localhost",
        user="root",
        password="",  # Replace with your password
        database="isatuamb_db"
    )
