backend/requirements.txt
Flask
flask-cors
requests
mysql-connector-python